Small improvements + move import/export logic to composables + add. version to export/import to handle future migrations
