+++ src/composables/useProjectExport.ts
@@ -1,7 +1,6 @@
 import useAsyncWebWorker from "@/composables/useAsyncWebWorker";
 import ProjectWorkerWebWorker from "./workers/projectWorker.worker?worker&inline";
 import {GroupAnalysisStore, GroupAnalysisStoreImport} from "@/store/new/GroupAnalysisStore";
-import JSZip from "jszip";
 
 export interface ProjectExportData {
     project: GroupAnalysisStoreImport;
